Refactor wait_if_busy and send into hardware trait
